Defined flag for swift version transition between v1 and v2
